LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-animation-dynamic-name-expected.txt


FAIL Animation with anchor() responds to anchor-name change assert_equals: expected 150 but got 200
